Add new parameter 'user_info_endpoint_url' to keycload_oidc
This patch adds the possibility to specify the endpoind which will be
used for autorization. Default value is
'/realms/%s/protocol/openid-connect/userinfo' where %s is a realm name,
that will be substituded in the middleware code.
The whole url is a concatenation of 'auth_url' and 'user_info_endpoint_url'
config parameters.

Change-Id: I769b1cb8102654e341a2bf4739c787622167267f