Fix creation of artifact with ':' in names.
Currently to prevent name/version colisions we look
for artifact with the same name/version before creation.
To do this we use simple 'list' operation. Unfortunately,
if name or version contains ':', it is considered as filter_op
delimeter and Glare returns 400 for these cases.

This code fixes this and adds explicit operator 'eq:' for each
value.

Change-Id: I7760019cdbe1f6e8c5bf1c3b22a3cf07392aff19
Closes-bug: #1697698