Implement scrubber util
Scrubber is a utility, that cleans up artifacts that have been marked
as deleted, when delayed_delete mechanism is activated. Enabling
delayed_delete option means that after delete API call, the artifact
will change its status to deleted, but it's data will not be deleted
immediately. Afterwards (according scrubber settings), scrubber will
detect the artifacts with 'deleted' statuses and delete them completely.

The scrubber can be run as a background periodically task or once as
a standalone tool.

Implements-blueprint: scrubber

Change-Id: Icada538521be927bf8e42512fa2d0be9c97c73ff