Optimize quota calculations when they are disabled
Currently glare calculates the amount of data and artifacts
even if the related quotas are disabled (value -1 is set).
It's not optimal, because additional requests to db are
required.

This code changes this and implements new behaviour when
requsts to bd are performed only if quotas are enabled.

Also it disables quotas by default.

Change-Id: Ie297aa1bc6f2be989bcc29353c8fec0101cf2e5b
(cherry picked from commit 8796021ca66dcbe31382253d38a78b874cc99dbe)