Revert scoped lock format
Recently scoped lock format was mistakenly changed
from "type_name:name:version" to "owner". New format
prevents simultaneuos modifications of different 
artifacts by one tenant, because it locks the whole
tenant scope, not just a single artifact.

For this reason it's suggested to revert the format as
it was before.

Closes-bug: #1720139

Change-Id: I9dc26da85f424a4ced1ca232d1f33c548b602f0f