Fix artifact init case.
In case artifact's field is defined to be
nullable=false and also doesn't have any
default, we had a problem:
While creating new artifact of this type
with a value for this field we got:

"BadRequest <artifact_name> field: <field_name>.
value: [None]. Exception.

That happened although a field value was provided.
This kind of error is expected if the field does
not get any value at creation, but not expected
if value is provided.

The following patch suggest a fix for that.
closes-bug: #1775502

Change-Id: I642249f067ab853b4b7d56a8da4db5809ee414c6