Add patch info to pre_upload_hook.
We got a specific case in which this change
is needed:

An aritifact type that only one instance is
allowed to be in 'active' status.

To enforce that we want to use pre_update_hook,
and check if there is any other artifact in 'active' status.
To do so we need json_patch info (we need to know if the json_patch
change status to "active" or "deactivated". 

This change aim to suuport this case and also support backward 
compatibility.


With this background, if the a user update 'active' artifact
to be with status "deactivated", we must know the patch info.

Change-Id: I6e85ff89144ca6d3a69c7a240de3f41bd1d999e7